+++ power_grid_model_c/power_grid_model/include/power_grid_model/component/link.hpp
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "branch.hpp"
+#include "edge.hpp"
 
 #include "../auxiliary/input.hpp"
 #include "../auxiliary/update.hpp"
@@ -13,21 +13,22 @@
 
 namespace power_grid_model {
 
-class Link final : public Branch {
+class Link final : public Edge {
   public:
     using InputType = LinkInput;
     using UpdateType = BranchUpdate;
     static constexpr char const* name = "link";
 
+    ComponentType math_model_type() const final { return ComponentType::link; }
+
     explicit Link(LinkInput const& link_input, double u1, double u2)
-        : Branch{link_input}, base_i_from_{base_power_3p / u1 / sqrt3}, base_i_to_{base_power_3p / u2 / sqrt3} {}
+        : Edge{link_input}, base_i_from_{base_power_3p / u1 / sqrt3}, base_i_to_{base_power_3p / u2 / sqrt3} {}
 
     // override getter
     double base_i_from() const override { return base_i_from_; }
     double base_i_to() const override { return base_i_to_; }
     double loading(double /* max_s */, double /* max_i */) const override { return 0.0; };
     double phase_shift() const override { return 0.0; }
-    bool is_param_mutable() const override { return false; }
 
   private:
     double base_i_from_;
