+++ power_grid_model_c/power_grid_model/include/power_grid_model/main_core/output.hpp
@@ -16,6 +16,7 @@
 #include "../component/branch.hpp"
 #include "../component/branch3.hpp"
 #include "../component/current_sensor.hpp"
+#include "../component/edge.hpp"
 #include "../component/fault.hpp"
 #include "../component/load_gen.hpp"
 #include "../component/node.hpp"
@@ -85,7 +86,7 @@ constexpr auto output_result(Component const& node, MainModelState<ComponentCont
 }
 
 // output branch
-template <std::derived_from<Branch> Component, steady_state_solver_output_type SolverOutputType>
+template <std::derived_from<Edge> Component, steady_state_solver_output_type SolverOutputType>
 constexpr auto output_result(Component const& branch, std::vector<SolverOutputType> const& solver_output,
                              Idx2D math_id) {
     using sym = decode_symmetry_v<SolverOutputType>;
@@ -95,7 +96,7 @@ constexpr auto output_result(Component const& branch, std::vector<SolverOutputTy
     }
     return branch.template get_output<sym>(solver_output[math_id.group].branch[math_id.pos]);
 }
-template <std::derived_from<Branch> Component, short_circuit_solver_output_type SolverOutputType>
+template <std::derived_from<Edge> Component, short_circuit_solver_output_type SolverOutputType>
 inline auto output_result(Component const& branch, std::vector<SolverOutputType> const& solver_output, Idx2D math_id) {
     if (math_id.group == disconnected) {
         return branch.get_null_sc_output();
