+++ power_grid_model_c/power_grid_model/include/power_grid_model/main_core/state_queries.hpp
@@ -11,6 +11,7 @@
 #include "../common/enum.hpp"
 #include "../component/branch.hpp"
 #include "../component/branch3.hpp"
+#include "../component/edge.hpp"
 #include "../component/node.hpp"
 #include "../component/regulator.hpp"
 #include "../component/shunt.hpp"
@@ -21,7 +22,7 @@
 #include <concepts>
 
 namespace power_grid_model::main_core {
-template <std::derived_from<Branch> ComponentType, class ComponentContainer>
+template <std::derived_from<Edge> ComponentType, class ComponentContainer>
     requires model_component_state_c<MainModelState, ComponentContainer, ComponentType>
 constexpr auto get_branch_nodes(MainModelState<ComponentContainer> const& state, Idx topology_sequence_idx) {
     return state.comp_topo->branch_node_idx[topology_sequence_idx];
@@ -55,7 +56,7 @@ constexpr auto get_math_id(MainModelState<ComponentContainer> const& state, Idx 
     return state.topo_comp_coup->node[topology_sequence_idx];
 }
 
-template <std::derived_from<Branch> ComponentType, class ComponentContainer>
+template <std::derived_from<Edge> ComponentType, class ComponentContainer>
     requires model_component_state_c<MainModelState, ComponentContainer, ComponentType>
 constexpr auto get_math_id(MainModelState<ComponentContainer> const& state, Idx topology_sequence_idx) {
     return state.topo_comp_coup->branch[topology_sequence_idx];
@@ -91,10 +92,10 @@ constexpr auto comp_base_sequence_cbegin(MainModelState<ComponentContainer> cons
     return state.reduced_topology->topo_node_coup.coupling.user_nodes_to_topo_nodes.cbegin();
 }
 
-template <std::derived_from<Branch> Component, class ComponentContainer>
+template <std::derived_from<Edge> Component, class ComponentContainer>
     requires model_component_state_c<MainModelState, ComponentContainer, Component>
 constexpr auto comp_base_sequence_cbegin(MainModelState<ComponentContainer> const& state) {
-    return state.topo_comp_coup->branch.cbegin() + get_component_sequence_offset<Branch, Component>(state.components);
+    return state.topo_comp_coup->branch.cbegin() + get_component_sequence_offset<Edge, Component>(state.components);
 }
 
 template <std::derived_from<Branch3> Component, class ComponentContainer>
